Refactor project for more sane structure
