+++ client/src/containers/DashboardContainer.js
@@ -1,0 +1,11 @@
+import React from 'react'
+
+import Dashboard from '../components/dashboard/Dashboard'
+
+export default class DashboardContainer extends React.Component {
+  render () {
+    return (
+      <Dashboard />
+    )
+  }
+}
