+++ client/src/containers/LandingPageContainer.js
@@ -1,0 +1,10 @@
+import React from 'react'
+import LandingPage from '../components/LandingPage'
+
+export default class LandingPageContainer extends React.Component {
+  render () {
+    return (
+      <LandingPage />
+    )
+  }
+}
