--- client/src/containers/Login.js
@@ -1,12 +1,0 @@
-import React from 'react'
-import LoginView from '../components/LoginView'
-
-export default class Login extends React.Component {
-  render () {
-    return (
-      <div>
-        <LoginView />
-      </div>
-    )
-  }
-}
