+++ client/src/containers/LoginContainer.js
@@ -1,0 +1,12 @@
+import React from 'react'
+import Login from '../components/Login'
+
+export default class LoginContainer extends React.Component {
+  render () {
+    return (
+      <div>
+        <Login />
+      </div>
+    )
+  }
+}
