+++ client/src/containers/LogoutContainer.js
@@ -2,7 +2,7 @@ import React from 'react'
 import Auth from '../modules/auth'
 import { Redirect } from 'react-router-dom'
 
-export default class Logout extends React.Component {
+export default class LogoutContainer extends React.Component {
   componentDidMount () {
     Auth.deauthenticateUser()
   }
