+++ client/src/containers/ProfileContainer.js
@@ -1,0 +1,10 @@
+import React from 'react'
+import Profile from '../components/profile/Profile'
+
+export default class ProfileContainer extends React.Component {
+  render () {
+    return (
+      <Profile />
+    )
+  }
+}
