+++ client/src/index.js
@@ -2,8 +2,8 @@ import React from 'react'
 import ReactDOM from 'react-dom'
 import { BrowserRouter } from 'react-router-dom'
 
-import './containers/main.css'
-import App from './containers/App'
+import './main.css'
+import App from './App'
 import registerServiceWorker from './registerServiceWorker'
 
 ReactDOM.render(
